requirements.txt
# Scraping
scrapy>=2.11
scrapy-playwright>=0.0.40
playwright>=1.40

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13

# Data processing
pandas>=2.1
rapidfuzz>=3.5
scikit-learn>=1.3
unidecode>=1.3

# Utilities
python-dotenv>=1.0
requests>=2.31
